perf(startup): lazy-load dashboard search
