+++ src/app/event-handlers.ts
@@ -163,6 +163,7 @@ class LazyUnifiedSettings implements UnifiedSettingsController {
 
 
 export interface EventHandlerCallbacks {
+  openSearch: (options?: { toggle?: boolean }) => void;
   updateSearchIndex: () => void;
   updateFlightSource?: (adsb: PositionSample[], military: MilitaryFlight[]) => void;
   loadAllData: () => Promise<void>;
@@ -199,6 +200,8 @@ export class EventHandlerManager implements AppModule {
   private boundMapWidthResizeMoveHandler: ((e: MouseEvent) => void) | null = null;
   private boundMapWidthEndResizeHandler: (() => void) | null = null;
   private boundMapFullscreenEscHandler: ((e: KeyboardEvent) => void) | null = null;
+  private searchButtonsInitialized = false;
+  private boundSearchKeyHandler: ((e: KeyboardEvent) => void) | null = null;
   private boundMobileMenuKeyHandler: ((e: KeyboardEvent) => void) | null = null;
   private boundPanelCloseHandler: ((e: Event) => void) | null = null;
   private boundWidgetModifyHandler: ((e: Event) => void) | null = null;
@@ -234,6 +237,7 @@ export class EventHandlerManager implements AppModule {
   }
 
   init(): void {
+    this.setupSearchControls();
     this.setupEventListeners();
     this.setupIdleDetection();
     this.setupTvMode();
@@ -409,6 +413,10 @@ export class EventHandlerManager implements AppModule {
       document.removeEventListener('keydown', this.boundMapFullscreenEscHandler);
       this.boundMapFullscreenEscHandler = null;
     }
+    if (this.boundSearchKeyHandler) {
+      document.removeEventListener('keydown', this.boundSearchKeyHandler);
+      this.boundSearchKeyHandler = null;
+    }
     if (this.boundMobileMenuKeyHandler) {
       document.removeEventListener('keydown', this.boundMobileMenuKeyHandler);
       this.boundMobileMenuKeyHandler = null;
@@ -449,24 +457,40 @@ export class EventHandlerManager implements AppModule {
     this.ctx.authModal = null;
   }
 
-  private setupEventListeners(): void {
+  setupSearchControls(): void {
     const openSearch = () => {
-      this.callbacks.updateSearchIndex();
-      this.ctx.searchModal?.open();
+      this.callbacks.openSearch();
     };
-    document.getElementById('searchBtn')?.addEventListener('click', () => {
-      track('search-open', { source: 'desktop' });
-      openSearch();
-    });
-    document.getElementById('mobileSearchBtn')?.addEventListener('click', () => {
-      track('search-open', { source: 'mobile' });
-      openSearch();
-    });
-    document.getElementById('searchMobileFab')?.addEventListener('click', () => {
-      track('search-open', { source: 'fab' });
-      openSearch();
-    });
+    if (!this.searchButtonsInitialized) {
+      const searchBtn = document.getElementById('searchBtn');
+      const mobileSearchBtn = document.getElementById('mobileSearchBtn');
+      const searchMobileFab = document.getElementById('searchMobileFab');
+      searchBtn?.addEventListener('click', () => {
+        track('search-open', { source: 'desktop' });
+        openSearch();
+      });
+      mobileSearchBtn?.addEventListener('click', () => {
+        track('search-open', { source: 'mobile' });
+        openSearch();
+      });
+      searchMobileFab?.addEventListener('click', () => {
+        track('search-open', { source: 'fab' });
+        openSearch();
+      });
+      this.searchButtonsInitialized = Boolean(searchBtn || mobileSearchBtn || searchMobileFab);
+    }
+    if (!this.boundSearchKeyHandler) {
+      this.boundSearchKeyHandler = (e: KeyboardEvent) => {
+        if ((e.metaKey || e.ctrlKey) && e.key.toLowerCase() === 'k') {
+          e.preventDefault();
+          this.callbacks.openSearch({ toggle: true });
+        }
+      };
+      document.addEventListener('keydown', this.boundSearchKeyHandler);
+    }
+  }
 
+  private setupEventListeners(): void {
     document.getElementById('copyLinkBtn')?.addEventListener('click', async () => {
       const shareUrl = this.getShareUrl();
       if (!shareUrl) return;
